refactor(auth): consolidate auth subcommands into single command package
Merge the separate login, logout, printaccesstoken and flag packages
into one internal/auth/command package, mirroring the structure used by
other domains. Subcommand constructors become unexported and the Auth()
constructor moves into command.go.
