+++ internal/auth/command/login.go
@@ -1,10 +1,10 @@
-package login
+package command
 
 import (
 	"context"
 	"os"
 
-	"github.com/nais/cli/internal/auth/flag"
+	"github.com/nais/cli/internal/auth/command/flag"
 	"github.com/nais/cli/internal/gcloud"
 	"github.com/nais/cli/internal/naisapi"
 	"github.com/nais/naistrix"
@@ -12,14 +12,8 @@ import (
 	"golang.org/x/term"
 )
 
-type loginFlags struct {
-	*flag.Auth
-	Nais bool `name:"nais" short:"n" usage:"Login using login.nais.io instead of gcloud."`
-	Yes  bool `name:"yes" short:"y" usage:"Automatically answer yes to all prompts."`
-}
-
-func Login(parentFlags *flag.Auth) *naistrix.Command {
-	flags := &loginFlags{Auth: parentFlags}
+func login(parentFlags *flag.Auth) *naistrix.Command {
+	flags := &flag.Login{Auth: parentFlags}
 	return &naistrix.Command{
 		Name:            "login",
 		Title:           "Log in to the Nais platform.",
