--- internal/auth/flag/flag.go
@@ -1,9 +1,0 @@
-package flag
-
-import (
-	"github.com/nais/cli/internal/flags"
-)
-
-type Auth struct {
-	*flags.GlobalFlags
-}
